requirements.txt
fastapi
Jinja2
uvicorn
gunicorn
requests
pytest
pre-commit
opencv-python
Pillow
onnx
onnxruntime
imageio
